chore: remove unused type guard functions from bitunixWs service
- Removed `isPriceData`, `isTickerData`, and `isDepthData` from `src/services/bitunixWs.ts`.
- Moved these functions to `src/tests/hardening/bitunix_ws.test.ts` to maintain test coverage.
- Consolidated `isTickerData` test case from `src/services/bitunixWs_hardening.test.ts` into `src/tests/hardening/bitunix_ws.test.ts`.
- Preserved `isTradeData` and `isSafe` in `src/services/bitunixWs.ts` as they are still in use.
