+++ src/tests/hardening/bitunix_ws.test.ts
@@ -17,7 +17,55 @@
 
 
 import { describe, it, expect } from "vitest";
-import { isPriceData, isTickerData, isDepthData } from "../../services/bitunixWs";
+
+// Legacy Type Guards (Moved from service for testing)
+const isSafe = (v: any) => {
+  if (typeof v === 'string') return true;
+  if (typeof v === 'number') return !isNaN(v) && isFinite(v);
+  return false;
+};
+
+function isPriceData(d: any): d is { fr?: any; nft?: any; lastPrice?: any; lp?: any; la?: any; ip?: any; } {
+  if (!d || typeof d !== 'object' || Array.isArray(d)) return false;
+
+  if (d.lastPrice !== undefined && !isSafe(d.lastPrice)) return false;
+  if (d.lp !== undefined && !isSafe(d.lp)) return false;
+  if (d.la !== undefined && !isSafe(d.la)) return false;
+  if (d.ip !== undefined && !isSafe(d.ip)) return false;
+  if (d.fr !== undefined && !isSafe(d.fr)) return false;
+
+  const hasSafePrice = (d.lastPrice !== undefined && isSafe(d.lastPrice)) ||
+                       (d.lp !== undefined && isSafe(d.lp)) ||
+                       (d.la !== undefined && isSafe(d.la)) ||
+                       (d.ip !== undefined && isSafe(d.ip));
+
+  const hasSafeFunding = (d.fr !== undefined && isSafe(d.fr));
+
+  return hasSafePrice || hasSafeFunding;
+}
+
+function isTickerData(d: any): d is {
+  volume?: any; v?: any; lastPrice?: any; close?: any;
+  high?: any; low?: any; quoteVolume?: any;
+  h?: any; l?: any; q?: any;
+} {
+  if (!d || typeof d !== 'object' || Array.isArray(d)) return false;
+
+  if (d.lastPrice !== undefined && !isSafe(d.lastPrice)) return false;
+  if (d.close !== undefined && !isSafe(d.close)) return false;
+  if (d.volume !== undefined && !isSafe(d.volume)) return false;
+
+  if (d.v !== undefined && !isSafe(d.v)) return false;
+  if (d.q !== undefined && !isSafe(d.q)) return false;
+  if (d.h !== undefined && !isSafe(d.h)) return false;
+  if (d.l !== undefined && !isSafe(d.l)) return false;
+
+  return (d.volume !== undefined || d.v !== undefined || d.lastPrice !== undefined || d.close !== undefined);
+}
+
+function isDepthData(d: any): d is { b: any[]; a: any[] } {
+  return d && Array.isArray(d.b) && Array.isArray(d.a);
+}
 
 describe("Bitunix WebSocket Fast Path Guards", () => {
     describe("isPriceData", () => {
@@ -61,6 +109,11 @@ describe("Bitunix WebSocket Fast Path Guards", () => {
         it("should reject invalid types", () => {
             expect(isTickerData({ lastPrice: {} })).toBe(false);
             expect(isTickerData({ volume: [] })).toBe(false);
+        });
+
+        it('should reject ticker with object values for "v"', () => {
+            const unsafe = { v: { malicious: true }, lastPrice: "100" };
+            expect(isTickerData(unsafe)).toBe(false);
         });
     });
